alloy-compact from block and header
